перенос пользователей в БД + рефакторинг пакета domain

+++ src/main/kotlin/ru/ac/uniyar/domain/Filters.kt
@@ -15,14 +15,14 @@ fun authenticationFilter(
         val requestWithUser = request.cookie("token")?.value?.let { token ->
             jwtTools.subject(token)
         }?.let { userId ->
-            fetchUserViaUserId(userId.toInt())
-            //fetchUserViaUserId(userId)
+            fetchUserViaUserId(userId)
         }?.let { user ->
             request.with( currentUser of user)
         }?: request
         next(requestWithUser)
     }
 }
+
 fun authorizationFilter(
     currentUser: RequestContextLens<User?>,
     permissionsLens: RequestContextLens<RolePermissions>,
